Moved fullscreen image overlay into modals module
