+++ src/components/modals/FullscreenImageOverlay/index.ts
