+++ src/components/modals/FullscreenImageOverlay/styles/imageControls.scss
