+++ src/components/modals/FullscreenImageOverlay/styles/imageInfo.scss
