+++ src/components/modals/FullscreenImageOverlay/styles/imageWrapper.scss
