+++ src/components/modals/FullscreenImageOverlay/styles/index.scss
