+++ src/components/modals/index.ts
@@ -1,3 +1,4 @@
 export * from './ActivationModal';
 export * from './ConfirmationModal';
+export * from './FullscreenImageOverlay';
 export * from './SubmissionModal';
