+++ src/components/molecules/SubmissionCard/SubmissionCard.stories.tsx
@@ -2,9 +2,9 @@ import { Meta, Story } from '@storybook/react';
 import React from 'react';
 import { RecoilRoot } from 'recoil';
 import { submissionData } from '../../../data';
+import { FullscreenImageOverlay } from '../../modals';
 import { CardList } from '../../organisms';
 import { VotingDragAndDropContext } from '../../providers';
-import { FullscreenImageOverlay } from '../../views';
 import SubmissionCard from './SubmissionCard';
 import { SubmissionCardProps } from './types';
