+++ src/components/views/index.ts
@@ -1,7 +1,6 @@
 export * from './CleverRedirectView';
 export * from './DashboardView';
 export * from './ErrorView';
-export { FullscreenImageOverlay } from './FullscreenImageOverlay';
 export * from './LoginView';
 export * from './ScheduleView';
 export * from './SignupView';
